+++ app/src/main/java/com/example/myapplication/MainActivitySemantics.kt
@@ -1,5 +1,5 @@
 package com.example.myapplication
 
 object MainActivitySemantics {
-    val greetingText: String = "greeting_text"
+    const val greetingText: String = "greeting_text"
 }
